+++ server/src/events/handlers/flight_landed_event_handler.py
@@ -13,7 +13,7 @@ from src.utils.enums import NotificationType
 
 class FlightLandedEventHandler(EventHandler):
     def handle(self, event: Event) -> None:
-        payload = EventPayloadSchema(**event.payload)
+        payload = EventPayloadSchema.model_validate(event.payload)
         session = SessionLocal()
 
         notifications: List[Notification] = []